requirements.txt
# ============================================================
# Core requirements — used in Docker (no PyTorch/CUDA)
# ============================================================

# Core
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Document parsing
PyMuPDF>=1.25.0
python-docx>=1.1.0

# LangChain
langchain>=1.0.0
langchain-core>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.4.0
langchain-text-splitters>=1.0.0

# Vector store
faiss-cpu>=1.9.0

# Hybrid retrieval (BM25 only — no torch dependency)
rank-bm25>=0.2.2

# OpenAI client
openai>=2.0.0

# Guardrails
nemoguardrails>=0.21.0

# API server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
slowapi>=0.1.9
python-multipart>=0.0.9
redis>=5.0.0
